Security and stability fixes
CVE FIXES:
- CVE-2016-4303: Fix buffer overflow in cJSON UTF-16 surrogate pair parsing
  * Add bounds checking with end_ptr for safe string parsing
  * Add unicode validation to prevent invalid character sequences
  * Set error pointer on validation failures for proper error handling

SECURITY IMPROVEMENTS:
- Fix timing side-channel in auth_key comparison (timing_sensitive_equals)
  * Remove early return on size mismatch that leaked key length
  * Use constant-time comparison with unified loop for all sizes
  * Compare against max length with zero-padding for constant time

- Fix non-constant-time string comparison in SCRAM authentication
  * Use crypto::compare_equal() instead of standard string comparison
  * Prevents timing attacks on password verification

- Strengthen SASLPrep implementation
  * Add ASCII-only validation to prevent Unicode equivalence attacks
  * Reject non-ASCII characters (0x80+) that require normalization
  * Reject ASCII control characters (0x00-0x1F, 0x7F)

- Increase PBKDF2 iterations for password migration
  * Change from 1 iteration to default_iteration_count (4096)
  * Improves security of migrated passwords

- Add warning for default empty admin password
  * Logs security warning when --initial-password is not specified

BUG FIXES:
- Fix null pointer dereferences from unchecked map::find() results:
  * perfmon_collection_repo.cc
  * namespace_interface_repository.cc
  * s2polygon.cc

- Fix unsigned integer underflow bugs:
  * changefeed.cc, control.cc, seq.cc: Use i + 1 < size pattern
  * calculate_contracts.cc: Add empty() check before subtraction
  * internal_node.cc: Add bounds check for offset calculation
  * js_pprint.cc: Change size_t to ssize_t for reverse loop

- Fix uninitialized variable bugs:
  * profile.hpp: Add in-class initializers for start_t, split_t, sample_t
  * term_storage.cc: Initialize info member in default constructor

- Fix signed/unsigned comparison mismatches:
  * parallel_traversal.cc: Use ssize_t for loop variable
  * leaf_node.cc: Use size_t for num_deleted

- Fix cJSON buffer safety:
  * Replace unsafe strcpy() with memcpy() using pre-calculated lengths
  * Add integer overflow checks before memory allocation
  * Replace sprintf() with snprintf() with proper bounds

CODE QUALITY:
- Fix iterator invalidation in changefeed.hpp truncate_top()
- Add safety documentation for self-deletion patterns in page.cc and data_block_manager.cc
- Suppress redundant --server-name warning when names match

All changes maintain backward compatibility while improving security
and stability.

+++ src/btree/leaf_node.cc
@@ -1666,7 +1666,7 @@ void erase_deletions(
     was pointing to a deletion entry that's been erased, we need to remove that entry
     from `pair_offsets`. */
     int src = 0, dst = 0;
-    int num_deleted = deletion_offsets.size();
+    size_t num_deleted = deletion_offsets.size();
     for (; src < node->num_pairs; ++src) {
         uint16_t off = node->pair_offsets[src];
         auto it = deletion_offsets.find(off);
